config: add AM_PROG_CC_C_O for per-target compilation flags
The use of per-target compilation flags with C sources requires
that the macro AM_PROG_CC_C_O be called from configure.ac

If the C compiler does not accept the -c and -o options simultaneously,
define NO_MINUS_C_MINUS_O. This macro actually tests both the compiler
found by AC_PROG_CC, and, if different, the first cc in the path.
The test fails if one fails. This macro was created for GNU Make
to choose the default C compilation rule.

Reviewed-by: Alan Coopersmith <alan.coopersmith@oracle.com>
Signed-off-by: Gaetan Nadon <memsize@videotron.ca>